+++ src/MultiplierReservationStation.java
@@ -13,8 +13,8 @@ public class MultiplierReservationStation extends ReservationStation{
             return str;
         }
         str += getOpcode() + "   ";
-        int val1 = getSourceValue1();
-        int val2 = getSourceValue2();
+        double val1 = getSourceValue1();
+        double val2 = getSourceValue2();
         str += val1 + "     " + val2 + "     " + getSourceReg1() + "    " + getSourceReg2() +  "    " + getMemoryAddress();
         return str;
     }
